app/flask/requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
opencv_python==4.10.0.84
pdf2image==1.17.0
Pillow==10.3.0
PyPDF2==3.0.1
imutils==0.5.4
numpy==2.0.0
scipy==1.13.1
flask-cors==4.0.1
